pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "inaturalist-downloader"
version = "0.1.0"
description = "Utilities for extracting fish species from iNaturalist and building image datasets."
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10",
]

[project.optional-dependencies]
yolo = [
  "ultralytics>=8",
]

[project.scripts]
inat-download = "inaturalist_downloader.commands.download:main"
inat-extract-species = "inaturalist_downloader.commands.extract_species:main"
inat-prepare-split = "inaturalist_downloader.commands.prepare_split:main"
inat-check-images = "inaturalist_downloader.commands.check_images:main"
inat-check-coverage = "inaturalist_downloader.commands.check_coverage:main"

[tool.setuptools.packages.find]
include = ["inaturalist_downloader*"]
